# src/swagger/v1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book API",
    "version": "1.0.0",
    "description": "API para gerenciar livros"
  },
  "paths": {
    "/books": {
      "post": {
        "summary": "Cria um novo livro",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Livro criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "Erro na requisição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao criar o livro"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Lista todos os livros",
        "responses": {
          "200": {
            "description": "Lista de livros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro na requisição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao listar os livros"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "O Senhor dos Anéis"
          },
          "author": {
            "type": "string",
            "example": "J.R.R. Tolkien"
          },
          "pages": {
            "type": "integer",
            "example": 1178
          },
          "status": {
            "type": "string",
            "example": "Disponível"
          }
        },
        "required": ["name", "author", "pages", "status"]
      }
    }
  }
}